Bugfix - Changing api-group for secrets in KE clusterrole
In this change we are changing the apigroup for secrets and configmaps in the Kube-Enforcer ClusterRole and ROle.
In order to validate its permission on these resources Kube-Enforcer checks for the api group "*". WIth the last
change the api group was changed to "", causing Kube-Enforcer to assume it does not have secrets permission in the
cluster. Hence it was not performing auto reistry integration.