Fix| Update kube-enforcer clusterrole for Kube-Bench (#582)
The latest Kube-Enforcer will run the latest version of Kube-Bench - v0.9.1 . In this Kube-Bench, certain checks have turned from Manual to Automated.
To run the the new automated checks, the Kube-Bench pod needs to list serviceaccounts in the cluster. Since the Kube-Bench pods share the same service-account as kube-enforcer
we are updating the service-account's permission to list seviceaccount resources . Otherwise Kube-Bench reports the result as action forbidden.